Added matching code and yellow orbs, fixed bugs, adjusted board gen
Orbs find their neighbors by raycasting around themselves. The thing is, when the board is generated the orbs that were placed can not find their neighbors because collisions can not be detected until the entire scene is ready.
So for now the generated orbs wait 1 second to find their neighbors.
When orbs find their neighbors, their neighbors will turn slightly clearer to show this.