Fixed bugs, added a few animations for clarity
So the bugs should, *SHOULD* be gone. Falling orbs are now handled by the board and not the orb that was last fired.  Also matching orbs will blink before they dissappear and falling orbs will shrink before they dissappear.
If the game stutters when making a big match thats most likely because the console is spamming lots of text for testing purposes.